Moved count.
